adding autoplay, starting on genetic algorithm to figure out optimal weights for autoplay

+++ src/state/tetris/action.rs
@@ -1,7 +1,7 @@
-#[derive(Debug, Clone, Copy, PartialEq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum Action {
-    Move(f32),
-    Rotate(f32),
+    Move(i32),
+    Rotate(i32),
     HardDrop,
     Reset,
     Hold,
